parser/requirements.txt
urllib3
BeautifulSoup4
psycopg2-binary
